requirements.txt
streamlit
langchain
langchain-text-splitters
faiss-cpu
pypdf
sentence-transformers
